pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "rest_serializers"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Accent Design Group Ltd", email = "support@accentdesign.co.uk" },
]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
]
dependencies = [
    'django',
    'djangorestframework',
]
dynamic = ["version", "description"]

[project.urls]
Homepage = "https://github.com/AccentDesign/Accent_RestSerializers"

[project.optional-dependencies]
test = [
    "black",
    "mypy",
    "ruff",
]

[tool.flit.sdist]
include = [
    "rest_serializers/"
]
exclude = [
    "scripts/",
    "tests/",
    "manage.py",
]

[tool.mypy]
ignore_missing_imports = true

[tool.black]
line-length = 88

[tool.ruff]
select = ["E", "F", "I"]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
combine-as-imports = true
